Refactoring faccioli
